WIP: Going to split batch_from into intialse and continue
