+++ executor/lib.rs
@@ -26,6 +26,8 @@ pub mod pipeline;
 pub(crate) mod reduce_executor;
 pub mod row;
 pub mod write;
+mod query_executor;
+mod program_executors;
 
 // TODO: use a bit-vec, since we have a continuously allocated range of positions
 // ---> for now, using a byte vec, which is 8x wasteful and on the heap!
